Fixed checkstyle errors. Changed name to VerifiableProducer. Added synchronization for thread safety on println statements.
